+++ EpicJewels/EffectHelpers/ItemDisplay.cs
@@ -1,4 +1,5 @@
 ﻿using EpicJewels.Common;
+using EpicJewels.GemEffects;
 using HarmonyLib;
 using Jewelcrafting;
 using System;
@@ -15,13 +16,27 @@ namespace EpicJewels.EffectHelpers
         [HarmonyPatch(typeof(ItemDrop.ItemData), nameof(ItemDrop.ItemData.GetTooltip), typeof(ItemDrop.ItemData), typeof(int), typeof(bool), typeof(float), typeof(int))]
         public static class ItemToolTipDisplayEnhancer
         {
+            public static List<ItemDrop.ItemData.ItemType> allowed_item_types = new() {
+                ItemDrop.ItemData.ItemType.Attach_Atgeir,
+                ItemDrop.ItemData.ItemType.Bow,
+                ItemDrop.ItemData.ItemType.OneHandedWeapon,
+                ItemDrop.ItemData.ItemType.TwoHandedWeapon,
+                ItemDrop.ItemData.ItemType.TwoHandedWeaponLeft
+            };
             private static void Postfix(ItemDrop.ItemData item, ref String __result)
             {
                 if (Player.m_localPlayer == null || Player.m_localPlayer.IsItemEquiped(item) == false || Config.EnableItemTooltipDisplay.Value == false) { return; }
+                if (!allowed_item_types.Contains(item.m_shared.m_itemType)) { return; }
 
                 Player.m_localPlayer.GetSkills().GetRandomSkillRange(out var min, out var max, item.m_shared.m_skillType);
                 float total_dmg = item.m_shared.m_damages.GetTotalDamage();
                 Dictionary<String, DmgModDetails> pdm_active = DetermineCharacterDamageModifiers();
+
+                float coinhoarder_power = Player.m_localPlayer.GetEffectPower<GemEffects.CoinHoarder.Config>("Coin Hoarder").Power;
+                if (coinhoarder_power > 0) {
+                    coinhoarder_power = CoinHoarder.CoinHoarderBonusCalc(Player.m_localPlayer);
+                }
+                float spellblade_power = Player.m_localPlayer.GetEffectPower<GemEffects.Spellsword.Config>("Spellsword").Power;
 
                 List<String> entry_lines = __result.Split('\n').ToList();
                 List<String> result_lines = new List<string>(entry_lines);
@@ -31,79 +46,79 @@ namespace EpicJewels.EffectHelpers
                     if (entry_lines[i].Length < 17) { continue; }
 
                     string line_desc = entry_lines[i].Split(':')[0].Trim();
-                    EpicJewels.EJLog.LogDebug($"ItemDescription {entry_lines[i]} | {line_desc}|");
+                    //EpicJewels.EJLog.LogDebug($"ItemDescription {entry_lines[i]} | {line_desc}|");
                     switch (line_desc)
                     {
                         case "$inventory_damage":
                             break;
                         case "$inventory_fire":
-                            if (pdm_active.ContainsKey("Coin Hoarder")) {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                            if (coinhoarder_power > 0 || spellblade_power > 0) {
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                         case "$inventory_poison":
-                            if (pdm_active.ContainsKey("Coin Hoarder")) {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                            if (coinhoarder_power > 0 || spellblade_power > 0) {
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                         case "$inventory_frost":
-                            if (pdm_active.ContainsKey("Coin Hoarder"))
+                            if (coinhoarder_power > 0 || spellblade_power > 0)
                             {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                         case "$inventory_blunt":
                             if (pdm_active.ContainsKey("Add Blunt Damage"))
                             {
                                 pdm_active["Add Blunt Damage"].added = true;
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, pdm_active["Add Blunt Damage"].power, pdm_active["Coin Hoarder"].power);
-                            } else if (pdm_active.ContainsKey("Coin Hoarder")) {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, pdm_active["Add Blunt Damage"].power, coinhoarder_power, spellblade_power);
+                            } else if (coinhoarder_power > 0 || spellblade_power > 0) {
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_blunt, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                         case "$inventory_slash":
                             if (pdm_active.ContainsKey("Add Slash Damage"))
                             {
                                 pdm_active["Add Slash Damage"].added = true;
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_slash, min, max, pdm_active["Add Slash Damage"].power, pdm_active["Coin Hoarder"].power);
-                            } else if (pdm_active.ContainsKey("Coin Hoarder")) {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_slash, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_slash, min, max, pdm_active["Add Slash Damage"].power, coinhoarder_power, spellblade_power);
+                            } else if (coinhoarder_power > 0 || spellblade_power > 0) {
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_slash, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                         case "$inventory_pierce":
                             if (pdm_active.ContainsKey("Add Pierce Damage"))
                             {
                                 pdm_active["Add Pierce Damage"].added = true;
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_pierce, min, max, pdm_active["Add Pierce Damage"].power, pdm_active["Coin Hoarder"].power);
-                            } else if (pdm_active.ContainsKey("Coin Hoarder")) {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_pierce, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_pierce, min, max, pdm_active["Add Pierce Damage"].power, coinhoarder_power, spellblade_power);
+                            } else if (coinhoarder_power > 0 || spellblade_power > 0) {
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_pierce, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                         case "$inventory_lightning":
                             if (pdm_active.ContainsKey("Add Lightning Damage"))
                             {
                                 pdm_active["Add Lightning Damage"].added = true;
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_lightning, min, max, pdm_active["Add Lightning Damage"].power, pdm_active["Coin Hoarder"].power);
-                            } else if (pdm_active.ContainsKey("Coin Hoarder")) {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_lightning, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_lightning, min, max, pdm_active["Add Lightning Damage"].power, coinhoarder_power, spellblade_power);
+                            } else if (coinhoarder_power > 0 || spellblade_power > 0) {
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_lightning, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                         case "$inventory_spirit":
                             if (pdm_active.ContainsKey("Add Spirit Damage"))
                             {
                                 pdm_active["Add Spirit Damage"].added = true;
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_spirit, min, max, pdm_active["Add Spirit Damage"].power, pdm_active["Coin Hoarder"].power);
-                            } else if (pdm_active.ContainsKey("Coin Hoarder")) {
-                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_spirit, min, max, 0f, pdm_active["Coin Hoarder"].power);
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_spirit, min, max, pdm_active["Add Spirit Damage"].power, coinhoarder_power, spellblade_power);
+                            } else if (coinhoarder_power > 0 || spellblade_power > 0) {
+                                result_lines[i] = AddModifierExplainer(entry_lines[i], total_dmg, item.m_shared.m_damages.m_spirit, min, max, 0f, coinhoarder_power, spellblade_power);
                             }
                             break;
                     }
                 }
 
-                // Add damage values that we are currently 
+                // Add damage values that provide extra bonuses, even if its not a default damage type to the weapon
                 foreach (var entry in pdm_active)
                 {
-                    if (entry.Value.added == true || entry.Key == "Coin Hoarder") { continue; }
+                    if (entry.Value.added == true) { continue; }
 
                     result_lines.Add(BuildModifierExplainer(entry.Value.localizedName, total_dmg, min, max, entry.Value.power));
                 }
@@ -145,14 +160,10 @@ namespace EpicJewels.EffectHelpers
             {
                 set_damage_modifiers.Add("Add Blunt Damage", new DmgModDetails() { added = false, localizedName = "$inventory_blunt", power = Player.m_localPlayer.GetEffectPower<GemEffects.AddBluntDamage.Config>("Add Blunt Damage").Power });
             }
-            if (Player.m_localPlayer.GetEffectPower<GemEffects.AddBluntDamage.Config>("Coin Hoarder").Power > 0)
-            {
-                set_damage_modifiers.Add("Coin Hoarder", new DmgModDetails() { added = false, power = GemEffects.CoinHoarder.CoinHoarderBonusCalc(Player.m_localPlayer) });
-            }
             return set_damage_modifiers;
         }
 
-        private static string AddModifierExplainer(string current_line, float total_dmg, float m_dmg_value, float min, float max, float bonus_power, float coinhoarder_mult = 1f)
+        private static string AddModifierExplainer(string current_line, float total_dmg, float m_dmg_value, float min, float max, float bonus_power, float coinhoarder_mult = 1f, float spellsword_mult = 0f)
         {
             float bonus_dmg = total_dmg * (bonus_power / 100);
             string[] line_arr = current_line.Split(' ');
@@ -162,13 +173,22 @@ namespace EpicJewels.EffectHelpers
             if (match.Success == false) {
                 return current_line;
             }
-            // EpicJewels.EJLog.LogDebug($"Match {match.Value}");
+            //EpicJewels.EJLog.LogDebug($"Match {match.Value}");
 
             int dmg;
             bool parsed = int.TryParse(match.Value, out dmg);
             if (parsed == false) {
                 return current_line; 
             }
+
+            if (spellsword_mult > 0) {
+                float eitr_cost = Player.m_localPlayer.GetEffectPower<GemEffects.EitrFused.Config>("Eitr Fused").Cost + 5;
+                if (Player.m_localPlayer.HaveEitr(eitr_cost)) {
+                    coinhoarder_mult += (1f + (spellsword_mult / 100f));
+                }
+            }
+
+
             line_arr[1] = $"<color=purple>{((dmg + bonus_dmg) * coinhoarder_mult).ToString("F1")}</color>";
             // Not sure if this will be more confusing or not, maybe just recoloring is enough
             // Add the sum of bonus damage